Refactor file downloads to support drawer download
