+++ www/src/acb_app.ts
@@ -1,4 +1,3 @@
-import JSZip from "jszip";
 import { Unit } from "./basic_utils.js";
 import { AcbAppRunMode, AppFunctionMode } from "./common/acb_app_types.js";
 import { fileBytesToString, loadFilesAsBytes } from "./file_reader.js";
@@ -6,71 +5,11 @@ import { FileEntry, FileKind, getFileManagerStore, modifyDrawerNotificationForUs
 import { loadTestFile } from "./debug.js";
 import { run_acb, run_acb_summary } from './pkg/acb_wasm.js';
 import { Result } from "./result.js";
-import { AppExportResultOk, AppResultOk, AppSummaryResultOk, FileContent, RenderTable } from "./acb_wasm_types.js";
+import { AppExportResultOk, AppResultOk, AppSummaryResultOk, RenderTable } from "./acb_wasm_types.js";
 import { getOutputStore, setAppFunctionViewMode } from "./vue/output_store.js";
 import { getAppInputStore, getSummaryDate } from "./vue/app_input_store.js";
 import { ErrorBox } from "./vue/error_box_store.js";
-import { asError } from "./http_utils.js";
-
-function makeZip(files: FileContent[]): Promise<Blob> {
-   return new Promise((resolve, reject) => {
-      try {
-         // Create a zip file from the file contents
-         const zip = new JSZip();
-         for (const file of files) {
-            zip.file(file.fileName, file.content);
-         }
-         zip.generateAsync({ type: "blob" })
-            .then(resolve)
-            .catch(reject);
-      } catch (error) {
-         reject(asError(error));
-      }
-   });
-}
-
-function makeFilenameDateString(): string {
-   let date_str = new Date().toISOString();
-   // Replace colons and dots for filename safety
-   date_str = date_str.replace(/[:.]/g, "-");
-   return date_str;
-}
-
-function downloadBlob(filename: string, blob: Blob) {
-   // Create a temporary link to trigger the download
-   const url = URL.createObjectURL(blob);
-   const a = document.createElement("a");
-   a.href = url;
-   a.style.display = "none";
-   a.download = filename;
-   document.body.appendChild(a);
-   a.click();
-   // Clean up the URL object
-   document.body.removeChild(a);
-   URL.revokeObjectURL(url);
-}
-
-function makeZipAndDownload(files: FileContent[]): void {
-   makeZip(files).then((zipBlob) => {
-      let date_str = makeFilenameDateString();
-      const filename = `acb_export_${date_str}.zip`;
-      downloadBlob(filename, zipBlob);
-   }).catch((err: unknown) => {
-      console.error("Error creating zip file: ", err);
-      ErrorBox.getMain().showWith({
-         title: "Export Error",
-         descPre: "An error occurred while creating the export zip file:",
-         errorText: String(err),
-      });
-   });
-}
-
-function downloadCsv(filenameBase: string, csvContent: string) {
-   let date_str = makeFilenameDateString();
-   const filename = `${filenameBase}_${date_str}.csv`;
-   const blob = new Blob([csvContent], { type: "text/csv" });
-   downloadBlob(filename, blob);
-}
+import { downloadCsv, makeZipAndDownload } from "./download_utils.js";
 
 class CommonRunOptions {
    constructor(
